LayoutTests/media/modern-media-controls/macos-compact-inline-media-controls/macos-compact-inline-media-controls-constructor.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>MacOSCompactInlineMediaControls</code> constructor.");

const mediaControls = new MacOSCompactInlineMediaControls;

shouldBeEqualToString("mediaControls.element.localName", "div");
shouldBeEqualToString("mediaControls.element.className", "media-controls mac inline compact");
shouldBe("mediaControls.layoutTraits", "LayoutTraits.macOS | LayoutTraits.Compact");

debug("");
debug("Controls bar");
shouldBe("mediaControls.controlsBar.children.length", "3");
shouldBeTrue("mediaControls.controlsBar.children[0] instanceof ButtonsContainer");
shouldBeTrue("mediaControls.controlsBar.children[1] instanceof ButtonsContainer");
shouldBeTrue("mediaControls.controlsBar.children[2] instanceof LayoutNode");

debug("");
debug("Left container");
shouldBe("mediaControls.controlsBar.children[0].buttons", "[mediaControls.playPauseButton, mediaControls.skipBackButton]");
shouldBeEqualToString("mediaControls.controlsBar.children[0].element.className", "buttons-container left");
shouldBe("mediaControls.controlsBar.children[0].leftMargin", "8");
shouldBe("mediaControls.controlsBar.children[0].rightMargin", "12");
shouldBe("mediaControls.controlsBar.children[0].buttonMargin", "12");

debug("");
debug("Right container");
shouldBe("mediaControls.controlsBar.children[1].buttons", "[mediaControls.muteButton, mediaControls.fullscreenButton]");
shouldBeEqualToString("mediaControls.controlsBar.children[1].element.className", "buttons-container right");
shouldBe("mediaControls.controlsBar.children[1].leftMargin", "12");
shouldBe("mediaControls.controlsBar.children[1].rightMargin", "8");
shouldBe("mediaControls.controlsBar.children[1].buttonMargin", "12");

debug("");
debug("Volume slider");
shouldBeEqualToString("mediaControls.controlsBar.children[2].element.className", "volume-slider-container");
shouldBe("mediaControls.controlsBar.children[2].children", "[mediaControls.volumeSlider]");
shouldBeFalse("mediaControls.controlsBar.children[2].visible");
shouldBe("mediaControls.volumeSlider.width", "60");

debug("");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
